requirements.txt
# requirements.txt
PyQt5>=5.15
Pillow>=9.0
requests>=2.28
flask>=2.2
flask-cors>=3.0
colorama>=0.4.6
yt-dlp>=2023.3.4
ffmpeg-python>=0.2.0
appdirs>=1.4
packaging>=21.3
pystray>=0.19
